Refactor SonarCloud configuration, update type imports, and add new user types
