+++ .github/workflows/main.yml
@@ -20,8 +20,6 @@ jobs:
         uses: pnpm/action-setup@v4
         with:
           version: 10
-          node-version: 20
-          cache: 'pnpm'
 
       - name: Install dependencies
         run: pnpm install --frozen-lockfile
@@ -59,8 +57,6 @@ jobs:
         uses: pnpm/action-setup@v4
         with:
           version: 10
-          node-version: 20
-          cache: 'pnpm'
 
       - name: Install pnpm
         run: npm install -g pnpm@latest-10
@@ -83,8 +79,6 @@ jobs:
         uses: pnpm/action-setup@v4
         with:
           version: 10
-          node-version: 20
-          cache: 'pnpm'
       
       - name: Install pnpm
         run: npm install -g pnpm@latest-10
@@ -100,15 +94,19 @@ jobs:
 
       - name: Run SonarQube Scan
         env:
-          FRONT_TOKEN: ${{ secrets.FRONT_TOKEN }}
+          SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
+          SONAR_HOST_URL: ${{ secrets.SONAR_HOST_URL }}
         run: |
           sonar-scanner \
             -Dsonar.projectKey=vanillmar_base-app \
             -Dsonar.organization=vanillmar \
-            -Dsonar.host.url=https://sonarcloud.io \
-            -Dsonar.token=$FRONT_TOKEN \
+            -Dsonar.host.url=$SONAR_HOST_URL \
+            -Dsonar.token=$SONAR_TOKEN \
             -Dsonar.sources=. \
             -Dsonar.javascript.lcov.reportPaths=./coverage/lcov.info
+            -Dsonar.test.inclusions=**/app/**,**/lib/**
+            -Dsonar.exclusions=**/*.d.ts, **/types/**, **/__mocks__/**
+            -Dsonar.coverage.exclusions=**/__tests__/**,**/hooks/**, **/components/**, **/public/**, **/node_modules/**, sonar.exclusions=**/*.d.ts, **/types/**, **/__mocks__/**
 
   build:
     name: Build Project
@@ -122,8 +120,6 @@ jobs:
         uses:  pnpm/action-setup@v4
         with:
           version: 10
-          node-version: 20
-          cache: 'pnpm'
 
       - name: Install dependencies
         run: pnpm install
